move theme switcher to seperate page
